refactor(live): extract startup recovery into LiveSessionRecoverer (#486)
Move the session/crash-recovery startup sequence out of LiveTradingEngine
into src/engines/live/recovery.py: balance recovery (active-session crash
path and 7-day clean-restart path, #668 invariants), persisted-position
reload with stale-OPEN self-healing (#657), idempotent risk-manager
re-registration, and startup exchange reconciliation (PositionReconciler
with the legacy SL-based fallback and its offline-fill bookkeeping).

The close-accounting helpers (_closed_base_quantity, _close_entry_fee_usd,
_close_position_portion) move to src/engines/live/trade_close_accounting.py,
shared by the exit path and the recovery fallback; trading_engine re-exports
them so existing imports (including tests) are unchanged.

Pure refactor: method bodies moved verbatim with self.->state. attribute
reads through a RecoveryEngineState Protocol that makes the full coupling
surface (including the four mutated attributes) explicit. The engine keeps
thin delegating wrappers. Deferred reconciliation imports are preserved so
existing patch targets keep working. Engine: 6,110 -> 5,368 lines.
