+++ src/engines/live/trading_engine.py
@@ -71,7 +71,16 @@ from src.engines.live.monitoring import (
     extract_ml_predictions,
     extract_sentiment_data,
 )
+from src.engines.live.recovery import LiveSessionRecoverer
 from src.engines.live.strategy_manager import StrategyManager
+
+# Re-exported close-accounting helpers: the exit path uses them directly and
+# tests import them from this module.
+from src.engines.live.trade_close_accounting import (
+    _close_entry_fee_usd,
+    _close_position_portion,
+    _closed_base_quantity,
+)
 from src.engines.shared.correlation_handler import CorrelationHandler
 from src.engines.shared.dynamic_risk_handler import DynamicRiskHandler
 from src.engines.shared.execution.execution_model import ExecutionModel
@@ -122,131 +131,6 @@ logger = logging.getLogger(__name__)
 Position = LivePosition
 # Trade uses BaseTrade which has all required fields plus MFE/MAE tracking
 Trade = BaseTrade
-
-
-def _closed_base_quantity(position: Position) -> float | None:
-    """Actual filled base-asset quantity represented by a closing ``Trade`` row.
-
-    ``position.quantity`` is the authoritative filled base quantity of the *original*
-    position (set from the entry fill, see LiveExecutionEngine.execute_entry). Partial
-    exits reduce ``current_size`` (a fraction of balance) but never mutate ``quantity``,
-    so the quantity actually being closed is ``position.quantity`` scaled by the fraction
-    of the original position remaining.
-
-    Returns ``None`` (so the Trade row stores NULL rather than a fabricated or negative
-    value) when the filled quantity is unknown or the sizing inputs are corrupt:
-    missing/non-positive/non-finite ``quantity``, ``original_size`` not finite-positive,
-    or ``current_size`` not finite-non-negative.
-    """
-    qty = getattr(position, "quantity", None)
-    if qty is None:
-        return None
-    try:
-        qty = float(qty)
-    except (TypeError, ValueError):
-        return None
-    if not math.isfinite(qty) or qty <= 0:
-        return None
-    # Fall back to ``size`` only when ``original_size`` was never set (None). An
-    # explicit 0 / negative / non-finite original_size is corrupt sizing state — return
-    # None rather than silently scaling against ``size`` and fabricating a quantity.
-    original = position.original_size if position.original_size is not None else position.size
-    current = position.current_size if position.current_size is not None else position.size
-    try:
-        original_f = float(original)
-        current_f = float(current)
-    except (TypeError, ValueError):
-        return None
-    if not (math.isfinite(original_f) and original_f > 0):
-        return None
-    if not (math.isfinite(current_f) and current_f > 0):
-        # current_size <= 0 is a degenerate/flat slice — store NULL rather than a fabricated
-        # 0.0 quantity (which _close_position_portion would still pair with the full fee).
-        return None
-    if current_f > original_f:
-        # Scale-ins grow current_size beyond original_size but do NOT update
-        # ``quantity`` (the original entry fill), so the held base quantity cannot
-        # be derived by scaling. Store NULL rather than over-report the fill.
-        return None
-    return qty * (current_f / original_f)
-
-
-def _close_entry_fee_usd(position: Position, execution_engine: Any) -> float:
-    """Entry commission (USD) for a closing ``Trade`` row's ``commission``.
-
-    Prefer the exact value booked at open (``position.metadata['entry_fee']``, the same
-    amount deducted from ``account_balances`` as the ``entry_fee_<symbol>`` event).
-
-    Positions recovered after a restart lose that metadata (the ``positions`` table does
-    not persist entry fee), so fall back to an **approximate** reconstruction: the
-    execution engine's fee model applied to the recovered entry notional. This equals the
-    booked fee exactly in the common case (entries booked at the modelled rate) and is a
-    close estimate otherwise (e.g. a live fill booked at the actual exchange commission) —
-    preferred over dropping the entry leg, which would understate ``trades.commission``
-    versus the ledger. Never raises; returns ``0.0`` when neither the metadata nor a
-    usable notional is available.
-    """
-    meta = getattr(position, "metadata", None) or {}
-    if "entry_fee" in meta:
-        try:
-            fee = float(meta["entry_fee"])
-            if math.isfinite(fee):
-                return fee
-        except (TypeError, ValueError):
-            pass
-    # Recovered position: reconstruct from persisted entry economics (quantity * entry
-    # price, falling back to size * entry_balance), via the same USD fee model.
-    try:
-        qty = getattr(position, "quantity", None)
-        entry_price = float(position.entry_price) if position.entry_price is not None else 0.0
-        if qty is not None and entry_price > 0:
-            notional = float(qty) * entry_price
-        elif position.entry_balance is not None and position.size is not None:
-            notional = float(position.entry_balance) * float(position.size)
-        else:
-            return 0.0
-        if notional > 0 and math.isfinite(notional):
-            fee = float(execution_engine.calculate_entry_fee(notional))
-            fee = fee if math.isfinite(fee) and fee >= 0 else 0.0
-            logger.info(
-                "Reconstructed approximate entry fee $%.4f for %s (no entry-fee "
-                "metadata; recovered position) from notional $%.2f",
-                fee,
-                getattr(position, "symbol", "?"),
-                notional,
-            )
-            return fee
-    except (AttributeError, TypeError, ValueError):
-        pass
-    return 0.0
-
-
-def _close_position_portion(position: Position) -> float:
-    """Fraction of the ORIGINAL position represented by this close
-    (``current_size`` / ``original_size``).
-
-    Used to scale the **entry** fee onto a closing ``Trade`` row so its
-    ``commission`` matches the row's portion-level ``size``/``quantity``/``pnl``
-    (a partially-exited position's final close is only the remaining slice). Returns
-    ``1.0`` — i.e. do not scale — for a full close (the common case), for missing or
-    corrupt sizing, and for scale-in state (``current_size > original_size``), so the
-    entry fee is never inflated.
-    """
-    original = position.original_size if position.original_size is not None else position.size
-    current = position.current_size if position.current_size is not None else position.size
-    try:
-        original_f = float(original)
-        current_f = float(current)
-    except (TypeError, ValueError):
-        return 1.0
-    if not (
-        math.isfinite(original_f) and original_f > 0 and math.isfinite(current_f) and current_f >= 0
-    ):
-        return 1.0
-    ratio = current_f / original_f
-    if ratio <= 0 or ratio > 1:
-        return 1.0
-    return ratio
 
 
 def _create_exchange_provider(provider: str, config: ConfigManager, testnet: bool = False):
@@ -922,6 +806,10 @@ class LiveTradingEngine:
 
         # Account monitor — snapshots, status lines, performance summaries.
         self.account_monitor = LiveAccountMonitor(engine_state=self)
+
+        # Startup recovery — session balance, persisted positions, exchange
+        # reconciliation. Reads/writes engine state at call time (#486).
+        self.session_recoverer = LiveSessionRecoverer(engine_state=self)
 
     def _apply_dynamic_risk_adjustment(
         self,
@@ -5078,661 +4966,31 @@ class LiveTradingEngine:
     def _recover_existing_session(self) -> float | None:
         """Try to recover balance from an existing session.
 
-        Prefers an active session (crash recovery); otherwise falls back to the
-        most recent matching session within 7 days (clean-restart path). When
-        there is genuinely no recent session to recover (new symbol/strategy,
-        fresh DB, or older than the window), returns None and the engine starts
-        fresh. Recovery runs unconditionally in every environment so staging
-        mirrors production — there is intentionally no bypass flag. (A
-        recovery-bypass env var silently orphaned open positions and reset to a
-        phantom balance; removed — see #668.)
+        See ``LiveSessionRecoverer.recover_existing_session`` for the full
+        crash-recovery / clean-restart semantics (#668).
         """
-        try:
-            # Prefer an active session (crash recovery path).
-            session_id = self.db_manager.get_active_session_id()
-            source = "active"
-
-            # Fallback: most recent matching session within 7 days (clean-restart path).
-            if session_id is None:
-                strategy = self._strategy_name()
-                session_id = self.db_manager.get_last_session_id(
-                    within_hours=168,  # 7 days — covers long-running paper trading sessions
-                    strategy_name=strategy,
-                    symbol=self._active_symbol,
-                )
-                source = "recent inactive"
-
-            if session_id is None:
-                logger.info("🆕 No recent session found, starting fresh")
-                return None
-
-            logger.info("🔍 Found %s session #%s", source, session_id)
-
-            # Clean restart (inactive session): remember it so start() can carry its
-            # OPEN positions forward into the new session — INDEPENDENT of whether a
-            # positive balance was recovered. A fully-liquidated session (balance 0)
-            # can still hold an OPEN position; gating this on balance > 0 would
-            # re-orphan it (#668, P2). The active/crash path reuses the session
-            # directly below and never needs this.
-            if source != "active":
-                self._recovered_inactive_session_id = session_id
-
-            recovered_balance = self.db_manager.recover_last_balance(session_id)
-            # Sanitize BEFORE the positivity filter below. recover_last_balance()
-            # can return a Decimal (Numeric column), and corrupt state can be
-            # non-finite — float(Decimal('Infinity'))/NaN -> inf/nan. The `> 0`
-            # check would otherwise silently drop -inf/NaN to None (engine starts
-            # on the default balance and may trade) or raise on Decimal('NaN').
-            # Coerce to a float invariant and fail fast on non-finite: a corrupt
-            # persisted balance must halt startup, never feed position sizing
-            # (CODE.md "Arithmetic & Financial Calculations").
-            if recovered_balance is not None:
-                recovered_balance = float(recovered_balance)
-                if not math.isfinite(recovered_balance):
-                    raise ValueError(
-                        f"Recovered balance is not finite ({recovered_balance!r}); "
-                        "refusing to start on corrupt persisted state."
-                    )
-            if recovered_balance and recovered_balance > 0:
-                # Crash recovery (active session): reuse the existing session ID so
-                # trades stay attributed to the same session row. Clean restarts create
-                # a new session below; their OPEN positions are carried forward via
-                # _recovered_inactive_session_id (set above).
-                if source == "active":
-                    self.trading_session_id = session_id
-                    # Register the reused session with the DB manager. create_trading_session
-                    # sets _current_session_id for NEW sessions, but this active-recovery path
-                    # reuses an existing one — without this, every session-scoped write that
-                    # falls back to _current_session_id (balance updates, etc.) fails with
-                    # "No active trading session" on the first trade after recovery (#41).
-                    self.db_manager.set_current_session(session_id)
-                    # Wire session context to execution engine so journaling works
-                    self.live_execution_engine.session_id = session_id
-                    self.live_execution_engine.strategy_name = self._strategy_name()
-                logger.info(
-                    "💾 Recovered balance $%.2f from %s session #%s",
-                    recovered_balance,
-                    source,
-                    session_id,
-                )
-                return recovered_balance
-
-            logger.warning("⚠️  Session #%s found but no balance to recover", session_id)
-            return None
-        except ValueError:
-            # Corrupt-balance invariant violation — must not be swallowed by the
-            # broad handler below (that would silently fall back to the default
-            # balance). Propagate so startup fails fast.
-            raise
-        except Exception as e:
-            logger.error("❌ Error recovering session: %s", e, exc_info=True)
-            return None
+        return self.session_recoverer.recover_existing_session()
 
     def _ensure_positions_registered_with_risk_manager(self) -> None:
-        """Register every tracked position with the risk manager.
+        """Register every tracked position with the risk manager (idempotent).
 
-        Idempotent. Re-registering a known position is a no-op for risk
-        managers that key on (symbol, side); for the few that count entries,
-        registering twice still leaves the position visible — strictly
-        better than the recovered-but-invisible state we are guarding
-        against.
-
-        Parity rationale:
-        - Backtest registers every position at entry
-          (src/engines/backtest/execution/entry_handler.py:407-421).
-        - Live's DB-recovery path registers
-          (src/engines/live/trading_engine.py:_recover_active_positions).
-        - Live's reconciler path (PositionReconciler._reconcile_filled_entry)
-          can also create positions via track_recovered_position but does
-          not register. This sweep closes that gap so per-symbol caps and
-          correlation gating see all active positions, matching the
-          invariant backtest assumes always holds.
+        See ``LiveSessionRecoverer.ensure_positions_registered_with_risk_manager``
+        for the parity rationale.
         """
-        if self.risk_manager is None:
-            return
-        try:
-            positions_snapshot = self.live_position_tracker.positions
-        except Exception as e:
-            logger.warning("Failed to snapshot positions for risk-manager sync: %s", e)
-            return
-
-        for position in positions_snapshot.values():
-            try:
-                # Use current_size (post-partial-exit) — passing the original
-                # ``size`` would silently re-inflate risk_manager.daily_risk_used
-                # on every re-registration, undoing the prior
-                # adjust_position_after_partial_exit. CODE.md "Position Fields"
-                # rules: ``current_size`` is the source of truth for capital
-                # currently deployed.
-                effective_size = (
-                    float(position.current_size)
-                    if position.current_size is not None
-                    else float(position.size)
-                )
-                if effective_size <= 0:
-                    # 100% partial-exit drained the position but the close
-                    # ack has not popped it from the tracker yet. Calling
-                    # ``update_position(size=0.0)`` would fail the size>0
-                    # validator and leave ``daily_risk_used`` inflated at
-                    # the original allocation. Drain the slot via
-                    # ``close_position`` so the next entry sees the right
-                    # remaining budget, then skip re-registration.
-                    try:
-                        self.risk_manager.close_position(position.symbol)
-                    except (KeyError, ValueError, AttributeError) as drain_err:
-                        logger.debug(
-                            "Risk manager close_position drain skipped for %s: %s",
-                            position.symbol,
-                            drain_err,
-                        )
-                    continue
-                self.risk_manager.update_position(
-                    symbol=position.symbol,
-                    # __post_init__ guarantees side is a PositionSide enum.
-                    side=cast(PositionSide, position.side).value,
-                    size=effective_size,
-                    entry_price=position.entry_price,
-                )
-            except Exception as e:
-                logger.warning(
-                    "Failed to register recovered position %s with risk manager: %s",
-                    position.symbol,
-                    e,
-                )
+        self.session_recoverer.ensure_positions_registered_with_risk_manager()
 
     def _recover_active_positions(self) -> None:
         """Recover active positions from database"""
-        try:
-            if not self.trading_session_id:
-                return
-
-            # Self-heal BEFORE reloading (works in paper too, no exchange calls):
-            # close any OPEN position in this session that already has a terminal
-            # Trade. Such rows are the #657 footgun — historically a closed Trade
-            # was logged without flipping positions.status, so the stale-OPEN row
-            # gets reloaded here with its old stop_loss and re-closed, producing a
-            # phantom duplicate trade. Healing first means get_active_positions
-            # below cannot return them. Belt-and-suspenders alongside the atomic
-            # status flip now done in log_trade.
-            try:
-                healed = self.db_manager.heal_positions_with_terminal_trades(
-                    self.trading_session_id
-                )
-                if healed:
-                    logger.info(
-                        "🩹 Self-healed %d stale-OPEN position(s) with terminal trades "
-                        "before recovery (session #%s)",
-                        healed,
-                        self.trading_session_id,
-                    )
-            except Exception as heal_err:
-                # Never let a heal failure block recovery; the atomic log_trade
-                # flip remains the primary defense going forward.
-                logger.warning(
-                    "Position self-heal failed before recovery (continuing): %s",
-                    heal_err,
-                )
-
-            # Get active positions from database
-            db_positions = self.db_manager.get_active_positions(self.trading_session_id)
-
-            if not db_positions:
-                logger.info("📊 No active positions to recover")
-                return
-
-            logger.info("🔄 Recovering %s active positions...", len(db_positions))
-
-            for pos_data in db_positions:
-                # Convert database position to Position object
-                # Handle both uppercase and lowercase side values from database
-                side_value = pos_data["side"]
-                if isinstance(side_value, str):
-                    side_value = side_value.lower()
-
-                stored_entry_balance = pos_data.get("entry_balance")
-                try:
-                    entry_balance = (
-                        float(stored_entry_balance)
-                        if stored_entry_balance is not None
-                        else float(self.current_balance)
-                    )
-                except (TypeError, ValueError):
-                    logger.warning(
-                        "Recovered position %s has invalid entry balance %s; falling back to current balance",
-                        pos_data.get("symbol"),
-                        stored_entry_balance,
-                    )
-                    entry_balance = float(self.current_balance)
-
-                # Resolve the tracker key: prefer entry_order_id (exchange),
-                # fall back to database ID string for backward compat
-                entry_order_id = pos_data.get("entry_order_id")
-                tracker_key = entry_order_id or str(pos_data["id"])
-
-                position = Position(
-                    symbol=pos_data["symbol"],
-                    side=PositionSide(side_value),
-                    size=pos_data["size"],
-                    entry_price=pos_data["entry_price"],
-                    entry_time=pos_data["entry_time"],
-                    entry_balance=entry_balance,
-                    stop_loss=pos_data.get("stop_loss"),
-                    take_profit=pos_data.get("take_profit"),
-                    unrealized_pnl=float(pos_data.get("unrealized_pnl", 0.0) or 0.0),
-                    unrealized_pnl_percent=float(
-                        pos_data.get("unrealized_pnl_percent", 0.0) or 0.0
-                    ),
-                    quantity=pos_data.get("quantity"),
-                    # Hydrate partial-operation state so a position partially exited
-                    # before a restart closes at its REMAINING size (and logs the
-                    # remaining base quantity), not the full original. Without this,
-                    # current_size/original_size fall back to size and the close
-                    # over-reports size, pnl fraction, and trades.quantity.
-                    original_size=pos_data.get("original_size"),
-                    current_size=pos_data.get("current_size"),
-                    partial_exits_taken=pos_data.get("partial_exits_taken", 0) or 0,
-                    scale_ins_taken=pos_data.get("scale_ins_taken", 0) or 0,
-                    last_partial_exit_price=pos_data.get("last_partial_exit_price"),
-                    last_scale_in_price=pos_data.get("last_scale_in_price"),
-                    order_id=tracker_key,  # Backward compat: used as _positions dict key
-                    tracker_key=tracker_key,
-                    exchange_order_id=entry_order_id,
-                    client_order_id=pos_data.get("client_order_id"),
-                    db_position_id=pos_data["id"],
-                    stop_loss_order_id=pos_data.get("stop_loss_order_id"),
-                )
-
-                # Validate recovered entry_price before tracking. Positions
-                # with invalid entry_price cannot be closed properly and would
-                # become orphaned in the tracker.
-                if position.entry_price <= 0 or not math.isfinite(position.entry_price):
-                    logger.critical(
-                        "SKIPPING recovery of position %s (%s): invalid entry_price %.8f. "
-                        "MANUAL RECONCILIATION REQUIRED.",
-                        position.symbol,
-                        position.order_id,
-                        position.entry_price,
-                    )
-                    continue
-
-                if position.order_id:
-                    self.live_position_tracker.track_recovered_position(
-                        position, db_id=pos_data.get("id")
-                    )
-
-                # Register recovered stop-loss order with OrderTracker for monitoring
-                if position.stop_loss_order_id and self.order_tracker:
-                    self.order_tracker.track_order(position.stop_loss_order_id, position.symbol)
-                    logger.info(
-                        f"📡 Recovered and tracking stop-loss order {position.stop_loss_order_id} "
-                        f"for position {position.symbol}"
-                    )
-
-                # Update risk manager tracking for recovered positions
-                if self.risk_manager:
-                    try:
-                        self.risk_manager.update_position(
-                            symbol=position.symbol,
-                            # __post_init__ guarantees side is a PositionSide enum.
-                            side=cast(PositionSide, position.side).value,
-                            size=position.size,
-                            entry_price=position.entry_price,
-                        )
-                    except Exception as e:
-                        logger.warning(
-                            "Failed to update risk manager for recovered position %s: %s",
-                            position.symbol,
-                            e,
-                        )
-
-                logger.info(
-                    "✅ Recovered position: %s %s @ $%.2f",
-                    pos_data["symbol"],
-                    pos_data["side"],
-                    pos_data["entry_price"],
-                )
-
-            logger.info("🎯 Successfully recovered %s positions", len(db_positions))
-
-        except Exception as e:
-            logger.error("❌ Error recovering positions: %s", e, exc_info=True)
+        self.session_recoverer.recover_active_positions()
 
     def _reconcile_positions_with_exchange(self) -> None:
+        """Reconcile local positions with exchange state on startup.
+
+        Delegates to PositionReconciler for comprehensive order-based verification
+        (via ``LiveSessionRecoverer``), falling back to legacy SL-based
+        reconciliation if the reconciler is unavailable.
         """
-        Reconcile local positions with exchange state on startup.
-
-        Delegates to PositionReconciler for comprehensive order-based verification.
-        Falls back to legacy SL-based reconciliation if reconciler unavailable.
-        """
-        if not self.exchange_interface or not self.enable_live_trading:
-            return
-
-        positions_snapshot = self.live_position_tracker.positions
-
-        # Run PositionReconciler regardless of position count — resolve_pending_orders
-        # must execute even when no positions were recovered (e.g. entry submitted but
-        # never persisted as a position before crash).
-        if self.trading_session_id:
-            try:
-                from src.engines.live.reconciliation import (
-                    PositionReconciler,
-                    Severity,
-                    run_orphaned_borrow_sweep,
-                )
-
-                use_margin = getattr(self.exchange_interface, "is_margin_mode", False)
-                reconciler = PositionReconciler(
-                    exchange_interface=self.exchange_interface,
-                    position_tracker=self.live_position_tracker,
-                    db_manager=self.db_manager,
-                    session_id=self.trading_session_id,
-                    max_position_size=self.max_position_size,
-                    use_margin=use_margin,
-                    fee_rate=self.live_execution_engine.fee_rate,
-                )
-
-                if not positions_snapshot:
-                    logger.info("📊 No local positions to reconcile — checking pending orders")
-                    results = reconciler.resolve_pending_orders()
-
-                    # After pending orders are resolved (so a fill that becomes a
-                    # position is adopted first), sweep any orphaned margin borrow.
-                    # No-op unless margin + flag enabled; safe when flat.
-                    if use_margin and self._active_symbol:
-                        run_orphaned_borrow_sweep(
-                            exchange=self.exchange_interface,
-                            position_tracker=self.live_position_tracker,
-                            db_manager=self.db_manager,
-                            session_id=self.trading_session_id,
-                            use_margin=use_margin,
-                            symbols=[self._active_symbol],
-                            cooldown_state=self._orphan_sweep_cooldown,
-                            lock_registry=self._base_asset_locks,
-                        )
-
-                    # Process results even with no positions — a filled entry
-                    # order may create a position, and critical issues must
-                    # still trigger close-only mode.
-                    critical_count = sum(1 for r in results if r.severity == Severity.CRITICAL)
-                    if critical_count > 0:
-                        logger.critical(
-                            "🚨 %d CRITICAL reconciliation issues — entering close-only mode",
-                            critical_count,
-                        )
-                        self._close_only_mode = True
-
-                    for r in results:
-                        if r.status == "corrected" and r.severity >= Severity.HIGH:
-                            for correction in r.corrections:
-                                logger.warning(
-                                    "⚠️ Auto-corrected %s #%s: %s",
-                                    r.entity_type,
-                                    r.entity_id,
-                                    correction.reason,
-                                )
-
-                    if results:
-                        corrections = sum(len(r.corrections) for r in results)
-                        logger.info(
-                            "✅ Pending order resolution complete: %d results, %d corrections, %d critical",
-                            len(results),
-                            corrections,
-                            critical_count,
-                        )
-                    return
-
-                logger.info("🔄 Reconciling %s positions with exchange...", len(positions_snapshot))
-                results = reconciler.reconcile_startup(positions_snapshot)
-
-                # Check for critical issues
-                critical_count = sum(1 for r in results if r.severity == Severity.CRITICAL)
-                if critical_count > 0:
-                    logger.critical(
-                        "🚨 %d CRITICAL reconciliation issues — entering close-only mode",
-                        critical_count,
-                    )
-                    # Route through the guarded helper so the CLOSE_ONLY event is
-                    # emitted on this startup-critical path too, not just runtime.
-                    self._enter_close_only_mode()
-
-                # Log HIGH severity auto-corrections (cancelled entries, SL fills)
-                for r in results:
-                    if r.status == "corrected" and r.severity >= Severity.HIGH:
-                        for correction in r.corrections:
-                            logger.warning(
-                                "⚠️ Auto-corrected %s #%s: %s",
-                                r.entity_type,
-                                r.entity_id,
-                                correction.reason,
-                            )
-
-                corrections = sum(len(r.corrections) for r in results)
-                logger.info(
-                    "✅ Reconciliation complete: %d results, %d corrections, %d critical",
-                    len(results),
-                    corrections,
-                    critical_count,
-                )
-                return
-            except Exception as e:
-                logger.warning(
-                    "PositionReconciler failed, falling back to legacy reconciliation: %s", e
-                )
-                self._record_event(
-                    EventType.ERROR,
-                    f"PositionReconciler failed, falling back to legacy reconciliation: {e}",
-                    severity="error",
-                    component="reconciler",
-                    error_code="RECONCILER_FALLBACK",
-                    exc=e,
-                )
-
-        # Legacy fallback: SL-based reconciliation (requires positions)
-        if not positions_snapshot:
-            logger.info("📊 No local positions to reconcile")
-            return
-
-        try:
-            positions_to_close = self.stop_loss_manager.find_offline_filled_stops(
-                positions_snapshot
-            )
-
-            # Close positions that were stopped out
-            for position, exit_price in positions_to_close:
-                logger.info(
-                    "🔄 Marking position %s as closed (stop-loss triggered offline)",
-                    position.symbol,
-                )
-                # Update balance based on stop-loss exit
-                if exit_price:
-                    fraction = (
-                        position.current_size
-                        if position.current_size is not None
-                        else position.size
-                    )
-                    # Guard against division by zero (pnl_percent handles this but we log)
-                    if position.entry_price <= 0:
-                        logger.error(
-                            "Invalid entry_price %s for position %s - skipping reconciliation",
-                            position.entry_price,
-                            position.symbol,
-                        )
-                        continue
-                    if exit_price <= 0 or not math.isfinite(exit_price):
-                        logger.error(
-                            f"Invalid exit_price {exit_price} for position "
-                            f"{position.symbol} - skipping reconciliation"
-                        )
-                        continue
-
-                    # Use shared pnl_percent for parity with backtest engine
-                    side_enum = Side.LONG if position.side == PositionSide.LONG else Side.SHORT
-                    pnl_pct_sized = pnl_percent(
-                        position.entry_price, exit_price, side_enum, fraction
-                    )
-
-                    # Use entry_balance for PnL calculation to maintain backtest-live parity
-                    basis_balance = (
-                        float(position.entry_balance)
-                        if position.entry_balance is not None and position.entry_balance > 0
-                        else self.current_balance
-                    )
-                    # Calculate exit fee for filled offline stop-loss
-                    # Slippage is zero for filled orders - slippage already occurred on exchange
-                    # and is reflected in the fill price. Matches execute_filled_exit behavior.
-                    exit_position_notional = (
-                        basis_balance * fraction * (exit_price / position.entry_price)
-                    )
-                    exit_fee = self.live_execution_engine.calculate_exit_fee(exit_position_notional)
-                    exit_slippage_cost = 0.0  # Slippage already in fill price
-                    # Calculate GROSS P&L for Trade.pnl (parity with backtest engine)
-                    # and NET P&L for balance updates
-                    gross_pnl = pnl_pct_sized * basis_balance
-                    realized_pnl = gross_pnl - exit_fee  # Net P&L for balance update
-
-                    # Deduct margin interest for short positions closed offline
-                    offline_interest_cost = 0.0
-                    if (
-                        getattr(self.exchange_interface, "is_margin_mode", False)
-                        and position.side == PositionSide.SHORT
-                    ):
-                        try:
-                            from src.engines.live.reconciliation import PositionReconciler
-
-                            tracker = MarginInterestTracker(self.exchange_interface)
-                            base_asset = PositionReconciler._extract_base_asset(position.symbol)
-                            interest_base = tracker.get_position_interest_cost(
-                                base_asset, position.entry_time
-                            )
-                            offline_interest_cost = interest_base * exit_price
-                            if offline_interest_cost > 0:
-                                realized_pnl -= offline_interest_cost
-                                logger.info(
-                                    "Deducted margin interest $%.4f from offline SL PnL for %s",
-                                    offline_interest_cost,
-                                    position.symbol,
-                                )
-                        except Exception as e:
-                            logger.warning(
-                                "Failed to query margin interest for offline SL %s: %s",
-                                position.symbol,
-                                e,
-                            )
-
-                    # Atomic balance update for offline stop-loss reconciliation
-                    if self.trading_session_id is not None:
-                        try:
-                            with self.db_manager.atomic_balance_update(
-                                balance_change=realized_pnl,
-                                reason=f"offline_stop_loss_{position.symbol}",
-                                updated_by="live_engine_reconciliation",
-                                correlation_id=position.order_id,
-                            ) as balance_result:
-                                self.current_balance = balance_result["new_balance"]
-                                logger.info(
-                                    f"💰 Adjusted balance for offline stop-loss: ${realized_pnl:+,.2f} "
-                                    f"(fee: ${exit_fee:.2f}) -> ${self.current_balance:,.2f}"
-                                )
-                        except Exception as balance_err:
-                            logger.error(
-                                "Failed to update balance for offline stop-loss %s: %s. Skipping reconciliation.",
-                                position.symbol,
-                                balance_err,
-                            )
-                            continue
-                    else:
-                        # No trading session - update balance directly
-                        self.current_balance += realized_pnl
-                        logger.info(
-                            f"💰 Adjusted balance for offline stop-loss: ${realized_pnl:+,.2f} "
-                            f"(fee: ${exit_fee:.2f}) -> ${self.current_balance:,.2f}"
-                        )
-                    # Store GROSS P&L in Trade.pnl for parity with backtest engine
-                    # Fees are tracked separately via performance_tracker.record_trade()
-                    trade = Trade(
-                        symbol=position.symbol,
-                        side=position.side,
-                        size=fraction,
-                        entry_price=position.entry_price,
-                        exit_price=exit_price,
-                        entry_time=position.entry_time,
-                        exit_time=datetime.now(UTC),
-                        pnl=gross_pnl,
-                        pnl_percent=pnl_pct_sized,
-                        exit_reason="stop_loss_offline",
-                    )
-                    self.performance_tracker.record_trade(
-                        trade=trade,
-                        fee=exit_fee + offline_interest_cost,
-                        slippage=exit_slippage_cost,
-                    )
-                    self.completed_trades.append(trade)
-                    if self.log_trades:
-                        self._log_trade(trade)
-
-                    # Persist trade to DB with margin interest cost
-                    if self.trading_session_id is not None:
-                        self.db_manager.log_trade(
-                            symbol=position.symbol,
-                            # __post_init__ guarantees side is a PositionSide enum.
-                            side=cast(PositionSide, position.side).value,
-                            entry_price=position.entry_price,
-                            exit_price=exit_price,
-                            size=fraction,
-                            pnl=gross_pnl,
-                            strategy_name=self._strategy_name(),
-                            exit_reason="stop_loss_offline",
-                            entry_time=position.entry_time,
-                            exit_time=datetime.now(UTC),
-                            session_id=self.trading_session_id,
-                            # Round-trip fee in USD: entry_fee (booked at open, or
-                            # reconstructed from the fee model for restart-recovered
-                            # positions; entry leg scaled to the closed portion) plus this
-                            # offline exit_fee. Same units as account_balances, where
-                            # realized_pnl above is already net of exit_fee.
-                            commission=(
-                                _close_entry_fee_usd(position, self.live_execution_engine)
-                                * _close_position_portion(position)
-                                + exit_fee
-                            ),
-                            quantity=_closed_base_quantity(position),
-                            margin_interest_cost=offline_interest_cost,
-                        )
-
-                # Stop tracking the SL order to prevent memory leak
-                if position.stop_loss_order_id and self.order_tracker:
-                    self.order_tracker.stop_tracking(position.stop_loss_order_id)
-
-                # Remove from local positions
-                if position.order_id:
-                    self.live_position_tracker.remove_position(position.order_id)
-                if self.risk_manager:
-                    try:
-                        self.risk_manager.close_position(position.symbol)
-                    except Exception as e:
-                        logger.warning(
-                            "Failed to update risk manager for reconciled position %s: %s",
-                            position.symbol,
-                            e,
-                        )
-
-                # Close in database
-                db_ids = self.live_position_tracker.position_db_ids
-                # Tracked positions always carry a non-None order_id.
-                position_db_id = db_ids.get(cast(str, position.order_id))
-                if position_db_id:
-                    self.db_manager.close_position(position_id=position_db_id)
-
-            if positions_to_close:
-                logger.info(
-                    f"🔄 Reconciliation complete: {len(positions_to_close)} positions "
-                    "closed (stopped out while offline)"
-                )
-            else:
-                logger.info("✅ All positions verified - no offline closures detected")
-
-        except Exception as e:
-            logger.error("❌ Error reconciling positions with exchange: %s", e, exc_info=True)
+        self.session_recoverer.reconcile_positions_with_exchange()
 
     def _handle_strategy_change(self, swap_data: dict[str, Any]):
         """Handle strategy change callback"""
